backend/pyproject.toml
[tool.poetry]
name = "astrasim-backend"
version = "0.1.0"
description = "FastAPI backend for the AstraSim event-driven Monte Carlo simulator."
authors = ["AstraSim Team <dev@astrasim.local>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
tenacity = "^9.0.0"
apscheduler = "^3.10.4"
asyncpg = "^0.29.0"
pydantic-settings = "^2.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
mypy = "^1.11.0"
ruff = "^0.6.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-ra"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["app"]
lint.select = ["E", "F", "I", "UP", "B", "C4", "SIM"]
lint.ignore = ["E203"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_equality = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

